% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno_10k}
\alias{geno_10k}
\title{Example genotype data}
\format{
\subsection{\code{geno_10k}}{

A data frame with 400 rows and 2 columns:
\describe{
\item{Sample_ID}{Sample ID}
\item{geno}{10k loci allele frequency genotype matrix}
...
}
}
}
\usage{
geno_10k
}
\description{
A 10k subset of allele frequency data from a larger simulated whole genome datasets
}
\keyword{datasets}
