% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{imputation_accuracy}
\alias{imputation_accuracy}
\title{Calculate Imputation Accuracy}
\usage{
imputation_accuracy(
  geno,
  prop_na = 0.2,
  repl = 1,
  k_neighbours = 10,
  l_loci = 10,
  cpus = 1,
  mem = 1,
  method = "ld_knni_fast"
)
}
\arguments{
\item{geno}{Either a matrix of allele frequencies with sample IDs as row names,
or a 2-column tibble with samples IDs in Col1 and a matrix-column
of allele frequencies in Col2. Samples across rows, Loci down columns.}

\item{prop_na}{The proportion of data to mask as missing in the dataset for
accuracy testing. Default is 0.2.}

\item{repl}{Number of replicate runs for imputation accuracy calculation.
Default is 1.}

\item{k_neighbours}{The number of k-neighbours (samples) to use in the KNN approach.}

\item{l_loci}{Number of loci, ordered based on correlation to the SNP to be imputed, to use when calculating nearest neighbours.}

\item{cpus}{The number of CPUs to use for parallel processing. Default is 1.}

\item{mem}{Memory availability in GiB. This is divided by cpus for per worker allocation (default: 1)}

\item{method}{Imputation method to use. Currently the options are "ld_knni" or "ld_knni_fast" (default: "ld_knni_fast"). When using the ld_knni_fast method missing data will be temporarily replaced with the mean loci allele frequency for correlation calculations (correlation calculation is faster on a complete matrix). This can reduce imputation accuracy, especially in datasets with high missing data, while the impact is smaller in datasets with lower levels of missing data.}
}
\value{
A tibble containing imputation statistics, currently raw bias (RB) for each replicate run.
}
\description{
This function calculates the imputation accuracy for missing data using
k-nearest neighbors (KNN) approach by masking a proportion of data as
missing. It can also be used to explore optimal k and l values
}
