% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute Missing Allele Frequencies}
\usage{
impute(
  geno,
  k_neighbours = 10,
  l_loci = 10,
  cpus = 1,
  mem = 1,
  method = "ld_knni_fast"
)
}
\arguments{
\item{geno}{Either a matrix of allele frequencies with sample IDs as row names,
or a 2-column tibble with samples IDs in Col1 and a matrix-column
of allele frequencies in Col2. Samples across rows, Loci down columns.}

\item{k_neighbours}{The number of k-neighbours (samples) to use in the KNN approach (default: 10).}

\item{l_loci}{Number of loci, ordered based on correlation to the SNP to be impute, to use when calculating nearest neighbours (default: 10).}

\item{cpus}{The number of CPUs to use for parallel processing (default: 1).}

\item{mem}{Memory availability in GiB. This is divided by cpus for per worker allocation (default: 1)}

\item{method}{Imputation method to use. Currently the options are "ld_knni" or "ld_knni_fast" (default: "ld_knni_fast"). When using the ld_knni_fast method missing data will be temporarily replaced with the mean loci allele frequency for correlation calculations (correlation calculation is faster on a complete matrix). This can reduce imputation accuracy, especially in datasets with high missing data, while the impact is smaller in datasets with lower levels of missing data.}
}
\value{
Depending on the input geno format, either a matrix of imputed allele
frequencies with sample IDs as the row names, or a 2-column tibble with
samples IDs in Col1 and a matrix-column of imputed allele frequencies
in Col2.
}
\description{
This function imputes missing allele frequency data using the
linkage disequilibrium k-nearest neighbour imputation algorithm
for allele frequencies.
}
