% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{matrix_knni}
\alias{matrix_knni}
\title{Imputation of a Matrix of Missing Allele Frequencies}
\usage{
matrix_knni(geno, k = 10, l = 6, cpus = 1, mem = 1, fast = FALSE)
}
\arguments{
\item{geno}{Matrix of allele frequency genotype data to be imputed.
Samples across rows, Loci down columns.}

\item{k}{Number of nearest neighbours (samples) to use.}

\item{l}{Number of closet loci to use.}

\item{cpus}{Number of cpus or threads to use for parallel processing.}

\item{mem}{Memory availability in GiB. This is divided by cpus for per worker allocation (default: 1).}

\item{fast}{TRUE/FALSE use a faster method but with an accuracy reduction.}
}
\value{
Matrix of allele frequency genotype data containing imputed
genotype frequencies.
}
\description{
This is an internal function, please use impute() for imputation.
This function uses knni logic to impute all missing allele frequency
datapoints withing a matrix.
}
