% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_freq_impute}
\alias{mean_freq_impute}
\title{Mean Allele Frequency Impute}
\usage{
mean_freq_impute(geno)
}
\arguments{
\item{geno}{Matrix of allele frequency genotype data.
Samples across rows, Loci down columns.}
}
\value{
Genotype matrix imputed using the basic mean allele frequency method.
}
\description{
Function to impute missing data with just the mean allele frequency at the locus
}
