% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_geno}
\alias{parse_geno}
\title{Parse Genotype Matrix}
\usage{
parse_geno(geno)
}
\arguments{
\item{geno}{Either a matrix of allele frequencies with sample IDs as row names,
or a 2-column tibble with samples IDs in Col1 and a matrix-column
of allele frequencies in Col2. Samples across rows, Loci down columns.}
}
\value{
A list containing a vector of sample names as the first element and a
genotype matrix as the second element.
}
\description{
Parse genotype matrix out of input format and store sample IDs if present.
}
