% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{sim_na_geno}
\alias{sim_na_geno}
\title{Simulated Missing Data}
\usage{
sim_na_geno(geno, prop_na = 0.2)
}
\arguments{
\item{geno}{Matrix of allele frequency genotype data. Samples across
rows, Loci down columns.}

\item{prop_na}{Proportion of missing data to be introduced, in addition
to any already existing}
}
\value{
Matrix of genotype data with missing data added as NAs.
}
\description{
This function masks a proportion of data as missing. Useful for imputation
accuracy testing purposes
}
