% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{singlepoint_knni}
\alias{singlepoint_knni}
\title{Imputation of a Single Missing Allele Frequency}
\usage{
singlepoint_knni(k = 20, l = 10, locus, geno, ld_order, ind, prop_l = 0.5)
}
\arguments{
\item{k}{Number of nearest (samples) to use.}

\item{l}{Number of closest loci to use.}

\item{locus}{Column index value for the locus that need imputing.}

\item{geno}{Matrix of allele frequency genotypes, SNPs in columns,
samples across rows.}

\item{ld_order}{Vector containing a ordered index from pairwise LD/
correlation values between locus and the geno matrix.}

\item{ind}{Row index val for the sample needing imputing.}

\item{prop_l}{due to missing data using l loci might not always be possible. prop_l is the the min prop of l loci required, or max non na loci will be used.}
}
\value{
An imputed genotype frequency.
}
\description{
This is an internal function, please use impute() for imputation.
This function uses knni logic to impute an allele frequency for a single
datapoint.
}
