% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_var}
\alias{weighted_var}
\title{Weighted Variance Calculation}
\usage{
weighted_var(x, w)
}
\arguments{
\item{x}{Vector of allele frequencies.}

\item{w}{Genotype distance value ^-1.}
}
\value{
Weighted variance
}
\description{
Weighted Variance Calculation
}
